Add --no-signin option for better system auth integration
Client.login() expects the "op signin" command to return a session token that
can be used for subsequent commands.  However, when 1Password CLI integration
with system authentication has been enabled, this token is not returned and the
helper errors out with `no session token found`.  In addition, the prompt for
the master password is redundant in this scenario.

To work around this problem, this commit adds a --no-signin flag that can be
passed to git-credential-1password to make it skip the login step.  The user
will be prompted for authentication by the operating system when needed.

Possibly closes: #20

Change-Id: I73902b27c474fdb83d5c202df6fb178f4a308d72